+++ libraries/base/changelog.md
@@ -106,6 +106,9 @@
 
   * Add new `displayException` method to `Exception` typeclass. (#9822)
 
+  * Add `Data.Bits.toIntegralSized`, a size-checked version of
+    `fromIntegral`. (#9816)
+
 ## 4.7.0.1  *Jul 2014*
 
   * Bundled with GHC 7.8.3